+++ resources/views/components/layouts/admin-form.blade.php
@@ -1,0 +1,15 @@
+<form wire:submit.prevent="save">
+
+    <div class="grid cols-75:25">
+
+        <div class="bx">
+            {{ $main }}
+        </div>
+
+        <div>
+            {{ $aside }}
+        </div>
+
+    </div>
+
+</form>