+++ resources/views/components/menus.blade.php
@@ -1,0 +1,17 @@
+<h1>Add menus!</h1>
+
+{{-- <nav {{ $attributes }}>
+
+    @foreach($file as $menuname => $linksArr)
+
+        @if ($withHeaders)
+        <div class="menu-title"> {{ Str::title($menuname) }}</div>
+        @endif
+
+       load menu
+
+    @endforeach
+
+    {{ $slot }}
+
+</nav> --}}