+++ resources/views/components/sidebar/main.blade.php
@@ -3,7 +3,7 @@
     @if(Route::has('login'))
 
         @auth
-            <x-gotime-menu menuname="main" filename="nav-user" class="menu" />
+            <x-gt-menu menuname="main" filename="nav-user" class="menu" />
         @else
             <div class="px">
                 {{-- use @include to prevent error if package not used --}}
@@ -15,7 +15,7 @@
 
     @endif
 
-    <x-gotime-menu menuname="main" class="menu" />
+    {{-- <x-gt-menu menuname="main" class="menu" /> --}}
 
     {{ $slot }}