+++ resources/views/components/toolbar/actions-toolbar.blade.php
@@ -5,28 +5,47 @@
 
 @aware(['routePrefix', 'editing'])
 
-@php
-    $resource = dotLastSegment($routePrefix)
-@endphp
+    @php
+        $resource = dotLastSegment($routePrefix)
+    @endphp
 
-<div id="actions-toolbar" class="pxy-05 light my flex space-between">
+    <div id="actions-toolbar" class="pxy-05 light my flex space-between">
 
-    <div>
+        <div>
 
-        <button wire:click.prevent="save('save_stay');" class="btn sm primary txt-upper">
-            <x-gt-icon-save class="icon" /> <span>Save</span> </button>
+            <button wire:click.prevent="save('save_stay');" class="btn primary txt-upper">
+                <x-gt-icon-save class="icon" /> <span>Save</span> </button>
 
-        <button wire:click.prevent="save('save_close')" class="btn sm primary txt-upper">
-            <x-gt-icon-exit class="icon" /> <span>Save and Close</span> </button>
+            <button wire:click.prevent="save('save_close')" class="btn primary txt-upper">
+                <x-gt-icon-exit class="icon" /> <span>Save and Close</span> </button>
+
+            <x-gt-button-delete wire:click.prevent="setConfirmAction({{ $editing->id }})" icon />
+
+        </div>
+
+        <div>
+
+            <a href="{{ route("$routePrefix.index") }}" class="btn txt-upper dark">
+                <x-gt-icon-exit class="icon" /> <span>{{ $resource }} Table</span>
+            </a>
+
+        </div>
 
     </div>
 
-    <div>
+    {{-- <x-gt-modal.confirmation wire:model="confirmingActionId">
 
-        <a href="{{ route("$routePrefix.index") }}" class="btn sm txt-upper dark">
-            <x-gt-icon-exit class="icon" /> <span>{{ $resource }} Table</span>
-        </a>
+        <x-slot name="title">
+            Delete Item
+        </x-slot>
 
-    </div>
+        <p> Are you sure you want to delete this item? </p>
 
-</div>
+        <x-slot name="footer">
+            <button wire:click.prevent="$set('confirmingActionId', false)"
+                wire:loading.attr="disabled" class="btn">Nevermind </button>
+            <button wire:click.prevent="delete({{ $confirmingActionId }})"
+                class="btn danger">Delete</button>
+        </x-slot>
+
+    </x-gt-modal.confirmation> --}}