+++ resources/views/layouts/partials/admin-nav.blade.php
@@ -1,0 +1,2 @@
+{{-- Septerate admin-nav to allow easy override --}}
+<x-gt-menu filename="nav-admin" class="menu" withIcons />