+++ resources/views/layouts/partials/navbar.blade.php
@@ -13,8 +13,7 @@ the container then publish the layout locally and override. --}}
 
         <div class="flex va-c hide-up-to-tablet">
 
-            {{-- main navigation --}}
-            <x-gotime-menu menuname="main" class="gg-3" itemClass="txt-white hover:txt-secondary" />
+            <x-gt-menu menuname="main" class="gg-3" itemClass="txt-white hover:txt-secondary" />
 
             @if(Route::has('checkout'))
                 <livewire:cart-button />